refactor tweak related to Audience moving from model.recommendation package to model.
